Formatted Game Details Page
The game titles were hidden behind the nav bar, adding some margins so everything displays properly. Also centered the back to game list button and gave it some margins as it was partially blocked by the footer.

removed the extra button in Admin page as well as the associated code in the Home Controller.